Actions: Add Linux build step (#654)
Related to https://github.com/apple/containerization/issues/606

To make sure the project (and many linux specific libraries) builds on
linux lets add a linux step to build for both musl and glibc. Eventually
this will run unit tests as well. This change needs a bit more libc
agnostic-ifying work to make sure things work for both libcs.

+++ vminitd/Sources/vmexec/RunCommand.swift
@@ -81,7 +81,7 @@ struct RunCommand: ParsableCommand {
         guard statfs("/", &s) == 0 else {
             throw App.Errno(stage: "statfs(/)")
         }
-        flags |= s.f_flags
+        flags |= UInt(s.f_flags)
 
         guard mount("", "/", "", flags, "") == 0 else {
             throw App.Errno(stage: "mount rootfs ro")
